requirements.txt
ocpp>=0.8
websocket>=8.0